crawlCCBY/requirements.txt
requests
python-dotenv
youtube-transcript-api
openai
fastapi
uvicorn
pydantic
